FaceChat/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
psycopg2-binary
motor
whisper
langchain
numpy
opencv-python
face-recognition
scikit-learn
google-generativeai
requests
python-dotenv
databases
asyncpg
deepface
